ai-service/requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
flask-cors==4.0.0
